+++ samples/bluetooth/mesh/light/CMakeLists.txt
@@ -9,6 +9,5 @@ project(NONE)
 
 target_sources(app PRIVATE
 	src/main.c
-	src/prov_handler.c
 	src/model_handler.c)
 target_include_directories(app PRIVATE include)
