+++ samples/bluetooth/mesh/light/prj.conf
@@ -12,9 +12,11 @@ CONFIG_FLASH_MAP=y
 CONFIG_FCB=y
 CONFIG_SETTINGS=y
 CONFIG_HWINFO=y
-CONFIG_DK_LIBRARY=y
 CONFIG_LOG=y
 CONFIG_LOG_PRINTK=y
+
+CONFIG_DK_LIBRARY=y
+CONFIG_BT_MESH_DK_PROV=y
 
 # Bluetooth configuration
 CONFIG_BT=y
