+++ subsys/bluetooth/mesh/CMakeLists.txt
@@ -29,3 +29,5 @@ zephyr_library_sources_ifdef(CONFIG_BT_MESH_LOC_CLI gen_loc_cli.c)
 
 zephyr_library_sources_ifdef(CONFIG_BT_MESH_PROP_SRV gen_prop_srv.c)
 zephyr_library_sources_ifdef(CONFIG_BT_MESH_PROP_CLI gen_prop_cli.c)
+
+zephyr_library_sources_ifdef(CONFIG_BT_MESH_DK_PROV dk_prov.c)
