+++ subsys/bluetooth/mesh/Kconfig.models
@@ -1,0 +1,156 @@
+#
+# Copyright (c) 2019 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-BSD-5-Clause-Nordic
+#
+
+menu "nRF BT Mesh Models"
+
+config BT_MESH_NRF_MODELS
+	bool
+	help
+	  Common Mesh model support modules, required by all Nordic BT Mesh
+	  models.
+
+
+config BT_MESH_ONOFF_SRV
+	bool "Generic OnOff Server"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic OnOff Server model.
+
+config BT_MESH_ONOFF_CLI
+	bool "Generic OnOff Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic OnOff Client model.
+
+
+config BT_MESH_LVL_SRV
+	bool "Generic Level Server"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Level Server model.
+
+config BT_MESH_LVL_CLI
+	bool "Generic Level Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Level Client model.
+
+
+menuconfig BT_MESH_DTT_SRV
+	bool "Generic Default Transition Time Server"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Default Transition Time Server model.
+
+if BT_MESH_DTT_SRV && BT_SETTINGS
+
+config BT_MESH_DTT_SRV_PERSISTENT
+	bool "Store the Default Transition Time persistently"
+	default y
+	help
+	  Enable persistent storage for the Default Transition Time state.
+
+endif
+
+config BT_MESH_DTT_CLI
+	bool "Generic Default Transition Time Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Default Transition Time Client model.
+
+
+config BT_MESH_PONOFF_SRV
+	bool "Generic Power OnOff Server"
+	select BT_MESH_NRF_MODELS
+	select BT_MESH_DTT_SRV
+	select BT_MESH_ONOFF_SRV
+	imply BT_SETTINGS
+	help
+	  Enable Mesh Generic Power OnOff Server model.
+
+config BT_MESH_PONOFF_CLI
+	bool "Generic Power OnOff Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Power OnOff Client model.
+
+
+config BT_MESH_PLVL_SRV
+	bool "Generic Power Level Server"
+	select BT_MESH_NRF_MODELS
+	select BT_MESH_LVL_SRV
+	select BT_MESH_PONOFF_SRV
+	help
+	  Enable Mesh Generic Power Level Server model.
+
+config BT_MESH_PLVL_CLI
+	bool "Generic Power Level Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Power Level Client model.
+
+
+config BT_MESH_BATTERY_SRV
+	bool "Generic Battery Server"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Battery Server model.
+
+config BT_MESH_BATTERY_CLI
+	bool "Generic Battery Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Battery Client model.
+
+
+config BT_MESH_LOC_SRV
+	bool "Generic Location Server"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Location Server model.
+
+config BT_MESH_LOC_CLI
+	bool "Generic Location Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Location Client model.
+
+config BT_MESH_PROP_SRV
+	bool "Generic Property Server"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Property Server models.
+
+config BT_MESH_PROP_CLI
+	bool "Generic Property Client"
+	select BT_MESH_NRF_MODELS
+	help
+	  Enable Mesh Generic Property Client models.
+
+menu "Generic Property Model configuration"
+
+config BT_MESH_PROP_MAXSIZE
+	int "Generic Property value max size"
+	default 8
+	range 1 376
+	help
+	  The upper boundary of a Generic Property value's size. The entire
+	  value with an 4 byte overhead must fit within a full TX packet
+	  payload (see BT_MESH_TX_SEG_MAX).
+
+config BT_MESH_PROP_MAXCOUNT
+	int "Generic Property value max count"
+	default 8
+	range 1 189
+	help
+	  The largest number of Generic Property values that can be listed by
+	  a Generic Property server. Each property ID is 2 bytes long, and
+	  the full list plus a 1 byte overhead must fit within a full TX packet
+	  payload (see BT_MESH_TX_SEG_MAX).
+
+endmenu
+
+endmenu
